NS-25: the logbook, natively
Dashboard, event detail, event form, track page and settings, on the typed
NavigationStack that replaces RootView's placeholder. One @Observable model per
screen owns its data and its writes; the views are layout, and nothing bypasses
APIClient — which *is* the offline layer, so there is no separate offline path to
forget to test.

Ported into the Kit with their JS test cases: LapStats (lap-stats.js) for the
per-session best-3 average, warmup and pace slope, and EventDates for the date
helpers in app.js plus fmtDate. EventDates keeps the JS's two different clocks
deliberately — todayISO in UTC, a picked date in the user's calendar — because at
10pm in New York those disagree by a day.

Deep links are parsed by a pure DeepLink in the Kit, so the routing table is
unit-tested rather than tapped through. /share/<slug> — the only pattern the
association file advertises — opens a read-only shared logbook; without it, tapping
someone's link showed *your* logbook. trackevolution://auth?code= is deliberately
not a route: ASWebAuthenticationSession consumes it, and racing it would burn a
one-time code.

Three bugs found on the way, each fixed with a test that would have caught it:

- Event.days was an Int against a REAL column. The web form steps by 0.5, so a
  1.5-day weekend didn't round — it threw, failing the decode of the whole events
  list for anyone who had logged a half day.
- OfflineStore's pending count lived only in memory, so it read zero on every
  launch. Everything that moves the queue is gated on it: writes made in the
  paddock sat in SQLite forever, the app stopped mentioning them, and a later write
  was sent directly, ahead of them.
- ProgressChart let Swift Charts pick the x domain, which anchors at zero. With
  Unix timestamps that meant an axis starting in 1970 and a season's progress
  squeezed into a vertical line.

Garage handling, as the spec asks to state explicitly: the dashboard's
maintenance-due strip and vehicle cards are **omitted**, not summarized. Wear
tracking is a deferred web-only feature, so Settings links out to the web app and a
UI test keeps the strip from creeping back. Absent for want of an endpoint rather
than by choice: editing a lap in place and reordering sessions — there is no
PUT /laps/:id and no sort route, and src/ is out of scope.

Verified against a real dev server: CoreScreensUITests walks all five screens and
OfflineWritesUITests repeats event/session/lap CRUD with the server unreachable,
then confirms the queue replays with temp ids substituted. Screenshots are attached
on success, because the charts and trackmap are Canvas-drawn and a bad render fails
no assertion — which is how the axis bug surfaced.

git diff --stat public/ src/ is empty.

Closes #88

+++ apps/ios/Packages/TrackEvolutionKit/Sources/TrackEvolutionKit/API/Drafts.swift
@@ -28,7 +28,8 @@ public struct EventDraft: Encodable, Hashable, Sendable {
     public var startDate: String
     public var trackName: String?
     public var trackId: Int?
-    public var days: Int?
+    /// Fractional — the column is `days REAL`. See `Event.days`.
+    public var days: Double?
     public var club: String?
     public var runGroup: String?
     public var car: String?
@@ -63,7 +64,7 @@ public struct EventPatch: Encodable, Hashable, Sendable {
     public var trackName: Patch<String> = .unchanged
     public var trackId: Patch<Int> = .unchanged
     public var startDate: Patch<String> = .unchanged
-    public var days: Patch<Int> = .unchanged
+    public var days: Patch<Double> = .unchanged
     public var club: Patch<String> = .unchanged
     public var runGroup: Patch<String> = .unchanged
     public var car: Patch<String> = .unchanged
